👩‍💻 Move shared cli interfaces to myst-cli
