+++ .changeset/hip-poets-flash.md
@@ -1,0 +1,6 @@
+---
+'myst-cli': patch
+'mystmd': patch
+---
+
+Move shared cli interfaces to myst-cli
