--- .changeset/tasty-pears-explode.md
@@ -1,5 +1,0 @@
----
-'mystmd': patch
----
-
-Buld mystmd as esm
