+++ packages/mystmd/package.json
@@ -5,8 +5,6 @@
   "author": "Rowan Cockett <rowan@curvenote.com>",
   "license": "MIT",
   "type": "module",
-  "exports": "./dist/index.js",
-  "types": "./dist/index.d.ts",
   "keywords": [
     "markdown",
     "latex",
@@ -39,9 +37,8 @@
     "lint:format": "npm run copy:version; prettier --check \"src/**/*.ts\"",
     "test": "npm run link; npm run copy:version; vitest run",
     "test:watch": "npm run link; npm run copy:version; vitest watch",
-    "build:esm": "tsc",
     "build:cli": "esbuild src/index.ts --bundle --outfile=dist/myst.cjs --platform=node --external:fsevents --target=node14",
-    "build": "npm-run-all -l clean copy:version -p build:cli build:esm"
+    "build": "npm-run-all -l clean copy:version -p build:cli"
   },
   "devDependencies": {
     "chalk": "^5.2.0",
