+++ packages/mystmd/src/site.ts
@@ -1,26 +1,6 @@
-import { Command } from 'commander';
-import { Session, startServer } from 'myst-cli';
+import type { Command } from 'commander';
+import { Session, startServer, makeStartCommand } from 'myst-cli';
 import { clirun } from './clirun.js';
-import {
-  makeKeepHostOption,
-  makeHeadlessOption,
-  makePortOption,
-  makeServerPortOption,
-  makeExecuteOption,
-  makeMaxSizeWebpOption,
-} from './options.js';
-
-export function makeStartCommand() {
-  const command = new Command('start')
-    .description('Start the current project as a website')
-    .addOption(makeExecuteOption('Execute Notebooks'))
-    .addOption(makeKeepHostOption())
-    .addOption(makeHeadlessOption())
-    .addOption(makePortOption())
-    .addOption(makeServerPortOption())
-    .addOption(makeMaxSizeWebpOption());
-  return command;
-}
 
 export function makeStartCLI(program: Command) {
   const command = makeStartCommand().action(clirun(Session, startServer, program));
