--- packages/mystmd/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig/base.json",
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "include": ["."],
-  "exclude": ["dist", "build", "node_modules", "src/**/*.spec.ts", "tests"]
-}
